fix bug with twitter tag finders and delete message errors
